Internal rename: namespaces, folders and files
