+++ src/back/CcrlExtensionsHost.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33424.131
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TlcvExtensionsHost", "TlcvExtensionsHost\TlcvExtensionsHost.csproj", "{CE888335-C05F-4147-A1DA-95B283B2821C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CcrlExtensionsHost", "CcrlExtensionsHost\CcrlExtensionsHost.csproj", "{CE888335-C05F-4147-A1DA-95B283B2821C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
