+++ src/back/CcrlExtensionsHost/Models/FenRequest.cs
@@ -1,3 +1,3 @@
-﻿namespace TlcvExtensionsHost.Models;
+﻿namespace CcrlExtensionsHost.Models;
 
 public record class FenRequest(string Fen);
