+++ src/back/CcrlExtensionsHost/appsettings.json
@@ -35,6 +35,6 @@
   "Serilog": {
     "MinimumLevel": {
       "Default": "Debug"
-      }
-}
+    }
+  }
 }
